+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [\[0.1.0\] - 2025-07-05](#010---2025-07-05)
 - [\[0.2.0\] - 2025-07-05](#020---2025-07-05)
+- [\[0.3.0\] - 2025-07-05](#030---2025-07-05)
 
 ## [0.1.0] - 2025-07-05
 
@@ -14,3 +15,8 @@
 - Add initial schema support
 - Add swagger documentation
 - Add swagger ui support at /docs
+
+## [0.3.0] - 2025-07-05
+
+- Simplify Typebox schema handling
+- Remove response standardizer